Shortcode/requirements.txt
asgiref==3.3.1
autopep8==1.5.4
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
djangorestframework==3.12.2
Faker==5.8.0
idna==2.10
mixer==7.1.2
psycopg2==2.8.6
pycodestyle==2.6.0
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.3
